backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
paho-mqtt==1.6.1
apscheduler==3.10.4
redis==5.0.1
ultralytics==8.0.231
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2+cpu
torchvision==0.16.2+cpu
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
